error fixed

+++ app/api/categories/route.ts
@@ -1,17 +1,25 @@
 import { Category } from "@/models/Category";
 import mongoose from "mongoose";
+import { isAdmin } from "../auth/[...nextauth]/route";
 
 export async function POST(req: Request) {
   mongoose.connect(process.env.MONGO_URL!);
   const { name } = await req.json();
-  const category = await Category.create({ name });
-  return Response.json(category);
+  if (await isAdmin()) {
+    const category = await Category.create({ name });
+    return Response.json(category);
+  } else {
+    return Response.json({});
+  }
 }
 
 export async function PUT(req: Request) {
   mongoose.connect(process.env.MONGO_URL!);
   const { _id, name } = await req.json();
-  await Category.updateOne({ _id }, { name });
+
+  if (await isAdmin()) {
+    await Category.updateOne({ _id }, { name });
+  }
   return Response.json(true);
 }
 export async function GET() {
@@ -23,6 +31,8 @@ export async function DELETE(req: Request) {
   mongoose.connect(process.env.MONGO_URL!);
   const url = new URL(req.url);
   const _id = url.searchParams.get("_id");
-  await Category.deleteOne({ _id });
+  if (await isAdmin()) {
+    await Category.deleteOne({ _id });
+  }
   return Response.json(true);
 }
